Added repository and changed ids to Long

+++ src/main/java/com/prueba/entrevista/c9c3c613304b/dtos/BrandDTO.java
@@ -1,5 +1,5 @@
 package com.prueba.entrevista.c9c3c613304b.dtos;
 
-public record BrandDTO(Integer id, String description) {
-    
+public record BrandDTO(Long id, String description) {
+
 }
