+++ src/main/java/com/prueba/entrevista/c9c3c613304b/dtos/ProductDTO.java
@@ -1,5 +1,5 @@
 package com.prueba.entrevista.c9c3c613304b.dtos;
 
-public record ProductDTO(Integer id, String description) {
-    
+public record ProductDTO(Long id, String description) {
+
 }
